Move test for fetch of batches to consumer tests
